Improve cutting methods in NodifyEditor
