+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 > - Breaking Changes:
 >	- Made the setter of NodifyEditor.IsPanning private
+>	- Renamed StartCutting to BeginCutting in NodifyEditor
 > - Features:
 >	- Added BeginPanning, UpdatePanning, EndPanning, CancelPanning and AllowPanningCancellation to NodifyEditor
+>	- Added UpdateCutting to NodifyEditor
 > - Bugfixes:
 	
 #### **Version 6.6.0**
